+++ orix/sampling/sampling_utils.py
@@ -70,7 +70,7 @@ conversion_dict = {
 }
 
 
-def create_equispaced_grid(resolution):
+def uniform_SO3_sample(resolution):
     """
     Returns rotations that are evenly spaced according to the Harr measure on
     SO3
